Feat(compiler/types): basic type definitions.

+++ limits/src/lib.rs
@@ -1,3 +1,3 @@
 #![doc = include_str!("../README.md")]
 
-pub type Size = usize;
+pub type Size = u64;
